compatibility fix (virtualenv, running test cases in py3)
