+++ setup.py
@@ -47,20 +47,7 @@ import sys
 import warnings
 from glob import glob
 
-
-def updatePyExec(bindir, executable=None):
-	"""Update fail2ban-python link to current python version (where f2b-modules located/installed)
-	"""
-	bindir = os.path.realpath(bindir)
-	if executable is None:
-		executable = sys.executable
-	pypath = os.path.join(bindir, 'fail2ban-python')
-	# if not exists or point to another version - update link:
-	isfile = os.path.isfile(pypath)
-	if not isfile or os.path.realpath(pypath) != os.path.realpath(executable):
-		if isfile:
-			os.unlink(pypath)
-		os.symlink(executable, pypath)
+from fail2ban.setup import updatePyExec
 
 
 # Wrapper to install python binding (to current python version):
@@ -77,6 +64,9 @@ class install_scripts_f2b(install_scripts):
 		updatePyExec(bindir)
 		return outputs
 
+
+# Update fail2ban-python env to current python version (where f2b-modules located/installed)
+updatePyExec(os.path.join(os.path.dirname(__file__), 'bin'))
 
 if setuptools and "test" in sys.argv:
 	import logging
